+++ libs/parsing/semantic-tokens/src/lib/get-semantic-tokens.ts
@@ -7,8 +7,8 @@ import {
 } from 'vscode-languageserver';
 
 import {
-  SEMANTIC_TOKEN_MODIFIERS,
-  SEMANTIC_TOKEN_TYPES,
+  SEMANTIC_TOKEN_MODIFIER_INDEX_LOOKUP,
+  SEMANTIC_TOKEN_TYPE_INDEX_LOOKUP,
 } from './semantic-tokens-config.interface';
 
 /**
@@ -50,8 +50,10 @@ export function GetSemanticTokens(parsed: IParsed): SemanticTokens {
             positions[z][0],
             positions[z][1],
             positions[z][2],
-            SEMANTIC_TOKEN_TYPES.indexOf(SemanticTokenTypes.class),
-            SEMANTIC_TOKEN_MODIFIERS.indexOf(SemanticTokenModifiers.definition)
+            SEMANTIC_TOKEN_TYPE_INDEX_LOOKUP[SemanticTokenTypes.class],
+            SEMANTIC_TOKEN_MODIFIER_INDEX_LOOKUP[
+              SemanticTokenModifiers.definition
+            ]
           );
         }
       }