+++ CHANGELOG.md
@@ -4,6 +4,26 @@ All notable changes to the "idl" extension will be documented in this file.
 
 For much more detail, see our [developer notes](./extension/docs/developer/dev-notes/README.md).
 
+## 3.0.4
+
+When `compile_opt` is not present with idl2, idl3, or strictarr, delineate between function calls and indexing with parentheses. It won't parse 100% correctly, but it allows us to identify and fix parentheses.
+
+- New error code (105 with alias "illegal-var-index")
+
+- Automatically gets fixed when formatting if you have auto-fix enabled (true by default)
+
+## 3.0.3
+
+Expose ability to set environment variables for the IDL process that appears in the debug console
+
+## 3.0.2
+
+Fixed a bug with documentation being generated when using the VSCode UI
+
+When executing batch files, wrap the path in quotes like compile
+
+When told to start a session of IDL, add button that will let you start IDL from the dialog that appears
+
 ## 3.0.1
 
 Fix badge URLs now that the extension is live